rafty v0.1.0 Rafty

An implementation of the Raft consensus algorithm written in pure Elixir. For learning purposes only and not production use.

Types

args()

args() :: %{
  server_name: server_name(),
  node_name: node_name(),
  fsm: atom(),
  log: atom()
}

The configuration of a server.

catch_exit_error()

catch_exit_error() :: {:error, :noproc | :timeout}

An error caught when a request exits.

catch_exit_ref_error()

catch_exit_ref_error() :: {:error, :noproc}

An error caught when a request exits with the associated reference to the request.

client_id()

client_id() :: non_neg_integer()

id()

id() :: {server_name(), node_name()}

A tuple {server_name, node_name} that represents the unique id of a server.

log_index()

log_index() :: non_neg_integer()

node_name()

node_name() :: atom()

The name of a node.

not_leader()

not_leader() :: {:not_leader, id() | nil}

An error returned when a server is not a leader. This error provides a hint to the leader, if any.

opt_id()

opt_id() :: id() | nil

A possibly nil id of a server.

rpc()

rpc() ::
  Rafty.RPC.AppendEntriesRequest.t()
  | Rafty.RPC.AppendEntriesResponse.t()
  | Rafty.RPC.RequestVoteRequest.t()
  | Rafty.RPC.RequestVoteResponse.t()

server_name()

server_name() :: atom()

The name of a server. Should be unique among all servers.

server_state()

server_state() :: :follower | :candidate | :leader

The state of server server. A server is either a follower, candidate, or a leader.

term_index()

term_index() :: non_neg_integer()

timestamp()

timestamp() :: integer()

Functions

execute(id, client_id, ref \\ make_ref(), payload, timeout \\ 5000)

execute(id(), client_id(), reference(), term(), timeout()) ::
  term() | not_leader() | catch_exit_ref_error()

Executes a command that may modify the Raft state machine on the server with the specified id. If the server is not a leader, this function will return the leader of the server, possibly nil. Clients must specify the client id created by register() when calling execute(). If execute() times out, it will return the reference associated with the execute call. If a client wants to retry the command, they must supply this reference in subsequent retries. Additionally, if a client is retrying a command, they must not make any execute commands with different references or the command may be executed twice.

leader(id, timeout \\ 5000)

leader(id(), timeout()) :: Rafty.id() | nil | catch_exit_error()

Returns the leader of the server with the specified id. If this function does not timeout, it will not return nil. If the server has no leader at the time of the call, it will wait until a leader has been elected before replying.

query(id, payload, timeout \\ 5000)

query(id(), term(), timeout()) :: term() | not_leader() | catch_exit_error()

Executies a read-only query of the Raft state machine on the server with the specified id. If the server is not a leader, this function will return the leader of the server, possibly nil. Any query should not mutate the Raft state machine because it does not persist in the Raft log, so server failures may cause Raft state machine inconsistencies.

register(id, timeout \\ 5000)

register(id(), timeout()) :: client_id()

Registers a client on the server with the specified id. If the server is not a leader, this function will return the leader of the server, possibly nil. Clients must call register() before calling execute().

start_server(args)

start_server(args()) :: DynamicSupervisor.on_start_child()

Starts a new server on the current server with the specified arguments.

status(id, timeout \\ 5000)

status(id(), timeout()) ::
  {server_state(), term_index(), log_index(), log_index()} | catch_exit_error()

Returns a tuple {server_state, term_index, commit_index, last_applied} of the server with the specified id.

terminate_server(id)

terminate_server(id()) :: :ok | {:error, :not_found}

Terminates a server with the specified id.