rafty v0.1.0 Rafty.FSM behaviour

A behavior that defines the finite state machine interface that the Raft protocol modifies and reads from.

Types

state()

state() :: term()

The state of the finite state machine.

Callbacks

execute(state, term)

execute(state(), term()) :: {state(), term()}

Executes a command that may modify the finite state machine with the specified state.

init()

init() :: state()

Initializes a finite state machine.

query(state, term)

query(state(), term()) :: term()

Executes a read-only query of the finite state machine with the specified state.