rafty v0.1.0 Rafty.Log behaviour

A behavior that defines the log interface that the Raft protocol uses to persist data.

Types

state()

state() :: term()

The state of the log.

Functions

merge_logs(old, new)

merge_logs([Rafty.Log.Entry.t()], [Rafty.Log.Entry.t()]) :: [
  Rafty.Log.Entry.t()
]

Merges two lists of logs together. All entries after the first conflict are removed from old and replaced with the entries from new. If there are no conflicts, and new is missing entries, then this function is a no-op.

Callbacks

append_entries(state, list, arg3)

append_entries(state(), [Rafty.Log.Entry.t()], Rafty.log_index()) :: state()

Appends a list of entries in the log with the specified state.

close(state)

close(state()) :: :ok

Closes the log with the specified state.

get_entries(state, arg2)

get_entries(state(), Rafty.log_index()) :: [Rafty.Log.Entry.t()]

Returns a list of entries in the log with the specified state.

get_entry(state, arg2)

get_entry(state(), Rafty.log_index()) :: Rafty.Log.Entry.t()

Returns an entry in the log with the specified state.

get_metadata(state)

get_metadata(state()) :: Rafty.Log.Metadata.t()

Returns the metadata recorded in the log with the specified state.

init(arg1)

init(Rafty.server_name()) :: state()

Initializes the log of the specified server.

length(state)

length(state()) :: non_neg_integer()

Returns the length of the log with the specified state.

set_metadata(state, arg2)

set_metadata(state(), Rafty.Log.Metadata.t()) :: state()

Sets of the metadata in the log with the specified state.