rafty v0.1.0 Rafty.Log.Server

A server that is the interface to the underlying Raft log.

Types

t()

t() :: %Rafty.Log.Server{log: module(), log_state: term()}

Functions

append_entries(arg, entries, index)

append_entries(Rafty.id(), [Rafty.Log.Entry.t()], Rafty.log_index()) :: :ok

Appends a list of entries to index of the specified server's Raft log.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

get_entries(arg, index)

get_entries(Rafty.id(), Rafty.log_index()) :: [Rafty.Log.Entry.t()]

Returns a list of entries from index to the end of the specified server's Raft log.

get_entry(arg, index)

get_entry(Rafty.id(), Rafty.log_index()) :: Rafty.Log.Entry.t() | nil

Returns the entry at index of the specified server's Raft log.

get_term_index(arg)

get_term_index(Rafty.id()) :: Rafty.term_index()

Returns the term of the specified server.

get_voted_for(arg)

get_voted_for(Rafty.id()) :: Rafty.id() | nil

Returns the server that the specified server voted for, if any.

increment_term_index(arg)

increment_term_index(Rafty.id()) :: Rafty.term_index()

Increments the term of the specified server.

length(arg)

length(Rafty.id()) :: non_neg_integer()

Returns the length of the specified server's Raft log.

name(server_name)

name(Rafty.server_name()) :: atom()

Returns the name of the server.

set_term_index(arg, term_index)

set_term_index(Rafty.id(), Rafty.term_index()) :: :ok

Sets the term of the specified server to term_index.

set_voted_for(arg, voted_for)

set_voted_for(Rafty.id(), Rafty.id() | nil) :: :ok

Sets the server that the specified serverd voted for, if any.

start_link(args)

start_link(Rafty.args()) :: {:ok, term()} | {:error, term()}

Starts a Rafty.Log.Server process linked to the current process.

stop(arg)

stop(Rafty.id()) :: :ok

Stops a Rafty.Log.Server process.