rafty v0.1.0 Rafty.Timer

A general-purpose timer that will send a message to self() after a specified duration has elapsed.

Types

t()

t() :: %Rafty.Timer{
  command: atom(),
  ref: reference() | nil,
  timer_ref: reference() | nil
}

Functions

new(command)

new(atom()) :: t()

Creates a new, unset timer that sends command to self() after a specified duration has elapsed.

reset(timer, timeout)

reset(t(), timeout()) :: t()

Resets the timer with a new timeout. The old timeout may still trigger a message, so you must check that the reference sent in the message is ref stored in the timer. If the references don't match, then the message must be ignored.

stop(timer)

stop(t()) :: t()

Stops the timer. The old timeout may still trigger a message, so you must check that the reference sent in the message is ref stored in the timer. If the references don't match, then the message must be ignored.